publish deals
