+++ integration_tests/src/tests/mod.rs
@@ -22,7 +22,8 @@ mod multisig_test;
 pub use multisig_test::*;
 mod power_scenario_tests;
 pub use power_scenario_tests::*;
-// mod publish_deals_test;
+mod publish_deals_test;
+pub use publish_deals_test::*;
 // mod replica_update_test;
 // mod terminate_test;
 // mod verified_claim_test;
